Rename client directory to src.
