+++ .eslintignore
@@ -1,2 +1,2 @@
 bundle.js
-client/external
+src/external
