+++ .travis.yml
@@ -15,7 +15,7 @@ before_script:
 
 script:
   - npm test
-  - npx eslint client __tests__
+  - npx eslint src __tests__
 
 after_script:
   - ./cc-test-reporter after-build --exit-code $TRAVIS_TEST_RESULT
