+++ src/components/GaugeComponent.js
