+++ src/components/WebSockStatusComponent.js
