+++ src/connection/Epics2WebPlugin.js
