+++ src/connection/MalcolmPlugin.js
