+++ src/connection/SimulatorPlugin.js
